Complete the graph generate step with edgemap output
